Update README.md
I removed the figure 1 from the paper, because it is very redundant with the "what if" figure, which was created two days earlier, and which was on German news.
I appreciate the filling-in of the data points, but it is also redundant with the other figure.
Moreover, the "delay" of 10 days is only the reporting delay. We need further time points for evidence accumulation.